+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Changelog
 
-## [Unreleased](https://github.com/kabamgamer/dd2-strategy-hub/tree/HEAD) (yyyy-mm-dd)
+## [v1.5.0](https://github.com/kabamgamer/dd2-strategy-hub/tree/v1.5.0) (2024-03-13)
 
 * Defense calculators: Added table view for defense lists
 * Defense calculators: Display damage type/elemental in defense stats
 * Defense calculators: Implemented shattered modifier in setups
 
-[Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.4.4...HEAD)
+[Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.4.4...v1.5.0)